# docs/frontend-backend/openapi-snapshot.yaml
openapi: 3.1.0
info:
  title: BigHead API
  version: 0.3.0
  description: 'Contrato canônico Sprint 3: matriz T01-T56 e implementação FastAPI.'
servers:
- url: http://localhost:8000
  description: Local
paths:
...
  /v1/knowledge/documents:
    get:
      tags:
      - knowledge
      summary: T35 - /v1/knowledge/documents
      operationId: t35Get
      security:
      - HTTPBearer: []
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: classification
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Classification
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      responses:
        '200':
          description: T35 successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeDocumentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          $ref: '#/components/responses/Problem'
      x-bighead-screen: T35
      x-bighead-role: member
      x-bighead-cache-event: lista paginada; `knowledge.documents.updated`
      x-bighead-critical-error: documento falho sem motivo
    post:
      tags:
      - knowledge
      summary: T36 - /v1/knowledge/documents
      operationId: t36Post
      security:
      - HTTPBearer: []
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 200
          pattern: .*\S.*
          title: Idempotency-Key
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KnowledgeUploadRequest'
      responses:
        '202':
          description: T36 successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeUploadResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          $ref: '#/components/responses/Problem'
      x-bighead-screen: T36
      x-bighead-role: member
      x-bighead-cache-event: job assinc; `knowledge.ingestion.updated`
      x-bighead-critical-error: parse error
  /v1/memory/items:
    get:
      tags:
      - memory
      summary: T37 - /v1/memory/items
      operationId: t37Get
      security:
      - HTTPBearer: []
      parameters:
      - name: kind
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Kind
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      responses:
        '200':
          description: T37 successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryItemResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          $ref: '#/components/responses/Problem'
      x-bighead-screen: T37
      x-bighead-role: member
      x-bighead-cache-event: refresh por contestacao; `memory.updated`
      x-bighead-critical-error: item expirado sem fonte
  /v1/search/semantic:
    post:
      tags:
      - search
      summary: T38 - /v1/search/semantic
      operationId: t38Post
      security:
      - HTTPBearer: []
      parameters:
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SemanticSearchRequest'
      responses:
        '200':
          description: T38 successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SemanticSearchResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          $ref: '#/components/responses/Problem'
      x-bighead-screen: T38
      x-bighead-role: member
      x-bighead-cache-event: '`no-store`; `search.semantic.executed`'
      x-bighead-critical-error: cross-tenant blocked
  /v1/crm/imports:
    post:
      tags:
      - crm
      summary: T39 - /v1/crm/imports
      operationId: t39Post
      security:
      - HTTPBearer: []
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 200
          pattern: .*\S.*
          title: Idempotency-Key
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrmImportRequest'
      responses:
        '202':
          description: T39 successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrmImportResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          $ref: '#/components/responses/Problem'
      x-bighead-screen: T39
      x-bighead-role: analyst/manager
      x-bighead-cache-event: job de importacao; `crm.import.updated`
      x-bighead-critical-error: duplicata sem preview de merge
  /v1/crm/leads:
    get:
      tags:
      - crm
      summary: T40 - /v1/crm/leads
      operationId: t40Get
      security:
      - HTTPBearer: []
      parameters:
      - name: stage
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Stage
      - name: ownerId
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Ownerid
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      responses:
        '200':
          description: T40 successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeadListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          $ref: '#/components/responses/Problem'
      x-bighead-screen: T40
      x-bighead-role: analyst/member
      x-bighead-cache-event: lista filtrada; `leads.updated`
      x-bighead-critical-error: lead sem owner
  /v1/crm/leads/{leadId}:
    get:
      tags:
      - crm
      summary: T41 - /v1/crm/leads/{leadId}
      operationId: t41Get
      security:
      - HTTPBearer: []
      parameters:
      - name: leadId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Leadid
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      responses:
        '200':
          description: T41 successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeadDetailResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '201':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
      x-bighead-screen: T41
      x-bighead-role: analyst/member
      x-bighead-cache-event: '`lead.follow_up.created`; retry idempotente'
      x-bighead-critical-error: lead indisponivel ou chave divergente
  /v1/crm/pipeline:
    get:
      tags:
      - crm
      summary: CRM opportunity pipeline board
      operationId: crmPipelineGet
      security:
      - HTTPBearer: []
      parameters:
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineBoardResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
      x-bighead-screen: T42
      x-bighead-role: manager
      x-bighead-cache-event: invalida board; `opportunities.updated`
      x-bighead-critical-error: campo obrigatorio faltante
  /v1/content/campaigns:
    get:
      tags:
      - content
      summary: T43 - /v1/content/campaigns
      operationId: t43Get
      security:
      - HTTPBearer: []
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: channel
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Channel
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      responses:
        '200':
          description: T43 successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          $ref: '#/components/responses/Problem'
      x-bighead-screen: T43
      x-bighead-role: analyst/member
      x-bighead-cache-event: cache curto; `campaigns.updated`
      x-bighead-critical-error: campanha pausada sem contexto
  /v1/content/assets:
    get:
      tags:
      - content
      summary: T44 - /v1/content/assets
      operationId: t44Get
      security:
      - HTTPBearer: []
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      responses:
        '200':
          description: T44 successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentAssetResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '201':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
      x-bighead-screen: T44
      x-bighead-role: member/reviewer
      x-bighead-cache-event: versao de ativo; `content.updated`
      x-bighead-critical-error: '`changes requested` sem diff'
    post:
      tags:
      - content
      summary: T44 - /v1/content/assets
      operationId: t44Post
      security:
      - HTTPBearer: []
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 200
          pattern: .*\S.*
          title: Idempotency-Key
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContentAssetCreateRequest'
      responses:
        '201':
          description: T44 successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentAssetResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '200':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
      x-bighead-screen: T44
      x-bighead-role: member/reviewer
      x-bighead-cache-event: versao de ativo; `content.updated`
      x-bighead-critical-error: '`changes requested` sem diff'
  /v1/content/publications/{id}/retry:
    post:
      tags:
      - content
      summary: T45 - /v1/content/publications/{id}/retry
      operationId: t45Post
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 200
          pattern: .*\S.*
          title: Idempotency-Key
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublicationRetryRequest'
      responses:
        '200':
          description: T45 successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicationRetryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
      x-bighead-screen: T45
      x-bighead-role: member
      x-bighead-cache-event: retry seguro; `publications.updated`
      x-bighead-critical-error: provider error
...
  /v1/crm/imports/{importId}/resume:
    post:
      tags:
      - crm
      summary: T39 - resume failed CRM import rows
      operationId: t39ResumePost
      security:
      - HTTPBearer: []
      parameters:
      - name: importId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Importid
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrmImportResumeRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrmImportResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/crm/accounts/{accountId}/merge:
    post:
      tags:
      - crm
      summary: T39 - merge duplicate CRM account
      operationId: t39AccountMergePost
      security:
      - HTTPBearer: []
      parameters:
      - name: accountId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Accountid
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrmAccountMergeRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrmAccountMergeResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/crm/opportunities/{id}/stage:
    post:
      tags:
      - crm
      summary: T42 - /v1/crm/opportunities/{id}/stage
      operationId: t42Post
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpportunityStageRequest'
      responses:
        '200':
          description: T42 successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpportunityStageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/crm/leads/{leadId}/follow-ups:
    post:
      tags:
      - crm
      summary: Create an idempotent lead follow-up
      operationId: crmLeadFollowUpPost
      security:
      - HTTPBearer: []
      parameters:
      - name: leadId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Leadid
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 200
          pattern: .*\S.*
          title: Idempotency-Key
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeadFollowUpRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeadFollowUpResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
  responses:
    Problem:
      description: Erro no formato RFC 9457
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    KnowledgeDocumentResponse:
      properties:
        documents:
          items:
            $ref: '#/components/schemas/KnowledgeDocument'
          type: array
          title: Documents
        counters:
          additionalProperties:
            type: integer
          type: object
          title: Counters
        nextCursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Nextcursor
      type: object
      required:
      - documents
      - counters
      title: KnowledgeDocumentResponse
    KnowledgeUploadRequest:
      properties:
        fileRef:
          type: string
          maxLength: 2000
          minLength: 1
          title: Fileref
        classification:
          type: string
          title: Classification
          default: medium
        visibility:
          type: string
          title: Visibility
          default: tenant
        title:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
          title: Title
      type: object
      required:
      - fileRef
      title: KnowledgeUploadRequest
...
    MemoryItemResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/MemoryItem'
          type: array
          title: Items
        sources:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Sources
        nextCursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Nextcursor
      type: object
      required:
      - items
      - sources
      title: MemoryItemResponse
...